Revert "Fix Register and login on the server side #10 (#11)"
This reverts commit c02ea0133b2407966dba67b164b8c65bae59ce83.
