--- server/.env.example
@@ -1,4 +1,0 @@
-PORT         = xxxx
-MONGO_DB_URI = "MONGO-URI-LINK"
-TOKEN_KEY    = "YOUR-SECRET-KEY"
-TOKEN_AGE    = 24 * 60 * 60 * 1000
