--- server/src/routes/authentication.router.js
@@ -1,10 +1,0 @@
-const express = require("express")
-const router = express.Router()
-
-const { login } = require("../authentication/login");
-const { register } = require("../authentication/register");
-
-router.route("/auth/login").post(login);
-router.route("/auth/register").post(register);
-
-module.exports = router;
