+++ src/helpers/utils.js
@@ -1,0 +1,5 @@
+
+exports.getDateNoTime = (date) => {
+  console.log(date.getDate())
+    return new Date(date.getFullYear(), date.getMonth(), date.getDate());
+  }
