+++ src/main.js
@@ -1,3 +1,5 @@
+// import runServer from "./server.js";
+
 const server = require('./server');
 
 const port = parseInt(process.env.PORT || 3000);
